requirements-dev.txt
django-debug-toolbar
django-nose
